// src/enums/rules.enum.ts
export enum RuleTypeEnum {
  INVENTORY_RULE = 'INVENTORY_RULE',
  PLATFORM_RULE = 'PLATFORM_RULE',
  SOV_RULE = 'SOV_RULE',
  STATE_CHANGE_RULE = 'STATE_CHANGE_RULE',
  BUDGET_RULE = 'BUDGET_RULE',
  BIDDING_STRATEGY_RULE = 'BIDDING_STRATEGY_RULE',
  KEYWORD_RULE = 'KEYWORD_RULE',
  TARGET_RULE = 'TARGET_RULE',
  BIDDER_RULE = 'BIDDER_RULE',
  UNKNOWN = 'UNKNOWN',
  KEYWORD_HARVESTING_RULE = 'KEYWORD_HARVESTING_RULE',
  PLACEMENT_RULE = 'PLACEMENT_RULE',
  PAGE_TYPE_RULE = 'PAGE_TYPE_RULE',
  DAYPARTING_RULE = 'DAYPARTING_RULE',
}

export enum RuleStatusEnum {
  ENABLED = 'ENABLED',
  PAUSED = 'PAUSED',
  ARCHIVED = 'ARCHIVED',
  DRAFT = 'DRAFT',
  ENDED = 'ENDED',
}

export enum RuleExecutionModeEnum {
  APPLY_CHANGES = 'APPLY_CHANGES',
  PREVIEW_ONLY = 'PREVIEW_ONLY',
}

export enum RuleEntityTypeIdEnum {
  CAMPAIGN_ID = 'CAMPAIGN_ID',
  ADGROUP_ID = 'ADGROUP_ID',
  ASIN = 'ASIN',
  KEYWORD_ID = 'KEYWORD_ID',
  TARGET_ID = 'TARGET_ID',
  AD_ITEM_ID = 'AD_ITEM_ID',
  ITEM_ID = 'ITEM_ID',
  SOURCE_TARGET_MAPPING_ID = 'SOURCE_TARGET_MAPPING_ID',
  PLACEMENT_CAMPAIGN_ID = 'PLACEMENT_CAMPAIGN_ID',
  PLATFORM_CAMPAIGN_ID = 'PLATFORM_CAMPAIGN_ID',
}
// ...
export enum AuditStatusEnum {
  IDLE = 'idle',
  PROCESSING = 'processing',
  SUCCESS = 'success',
  WARNING = 'warning',
}
// ...
export enum RuleDetailsTypeEnum {
  RULE = 'rule',
  TEMPLATE = 'template',
}